+++ src/base/b-list/b-list.ts
@@ -208,6 +208,13 @@ class bList extends iData implements iVisible, iWidth, iActiveItems {
 
 	protected itemsStore!: this['Items'];
 
+	/**
+	 * If true, then item attribute `data-router-prevent-transition` will be set to `true`.
+	 * The field can be provided as a function. In that case, a value is taken from the result of invoking.
+	 */
+	@system()
+	protected routerPreventTransition?: boolean | ((item: Item) => boolean);
+
 	/** @see [[iActiveItems.prototype.activeElement] */
 	@computed({
 		cache: true,
@@ -505,6 +512,26 @@ class bList extends iData implements iVisible, iWidth, iActiveItems {
 	/** @see [[iItems.getItemKey]] */
 	protected getItemKey(item: this['Item'], i: number): CanUndef<IterationKey> {
 		return iItems.getItemKey(this, item, i);
+	}
+
+	/**
+	 * Returns href for item. Is used at the rendering stage
+	 *
+	 * @param item
+	 */
+	protected getHref(item: Item): CanUndef<string> {
+		return item.href;
+	}
+
+	/**
+	 * Returns true if it should prevent the router transition
+	 *
+	 * @param item
+	 */
+	protected getRouterPreventTransition(item: Item): boolean {
+		return Object.isFunction(this.routerPreventTransition) ?
+			this.routerPreventTransition(item) :
+			this.routerPreventTransition ?? false;
 	}
 
 	protected override initModEvents(): void {